PS-853_resilient-phraseanet-indexer (#588)
- import empty stories then records ; import stories while importing record (needs PHRAS-4151_stories-reference-in-api).
- track last story/record imported into .lock file(s)
- dump full phraseanet equivalent conf if `renditions` or `fieldMap` is not defined in conf, then stops (bc break).
